+++ R/count_n_tmhs.R
@@ -3,9 +3,9 @@
 #' @examples
 #' count_n_tmhs("00000000000000000000000000")
 #' count_n_tmhs("00000000001111100000000000")
+#' count_n_tmhs(c("0", "1"))
 #' @export
-count_n_tmhs <- function(topology_str) {
-  n_tmhs <- stringr::str_count(topology_str, "01")
-  if (substring(topology_str, 1, 1) == "1") n_tmhs <- n_tmhs + 1
-  n_tmhs
+count_n_tmhs <- function(topology_strs) {
+  stringr::str_count(topology_strs, "01") +
+    (substring(topology_strs, 1, 1) == "1")
 }